requirements.txt
python-telegram-bot==22.6
aiohttp==3.8.5
python-dotenv==1.0.0